+++ docs/cog_guides/modlog.rst
@@ -19,7 +19,7 @@ find detailed docs about usage and commands.
 Usage
 -----
 
-Manage log channels for moderation actions.
+Browse and manage modlog cases.
 
 
 .. _modlog-commands:
@@ -87,82 +87,6 @@ List cases for the specified member.
 **Arguments**
 
 * ``<member>``: The member to get cases for. |member-input|
-
-.. _modlog-command-modlogset:
-
-^^^^^^^^^
-modlogset
-^^^^^^^^^
-
-.. note:: |guildowner-lock|
-
-**Syntax**
-
-.. code-block:: none
-
-    [p]modlogset 
-
-**Description**
-
-Manage modlog settings.
-
-.. _modlog-command-modlogset-cases:
-
-"""""""""""""""
-modlogset cases
-"""""""""""""""
-
-**Syntax**
-
-.. code-block:: none
-
-    [p]modlogset cases [action]
-
-**Description**
-
-Enable or disable case creation for a mod action.
-
-**Arguments**
-
-* ``[action]``: The action to enable or disable case creation for.
-
-.. _modlog-command-modlogset-modlog:
-
-""""""""""""""""
-modlogset modlog
-""""""""""""""""
-
-**Syntax**
-
-.. code-block:: none
-
-    [p]modlogset modlog [channel]
-
-.. tip:: Alias: ``modlogset channel``
-
-**Description**
-
-Set a channel as the modlog.
-
-**Arguments**
-
-* ``[channel]``: The channel to set as the modlog. If omitted, the modlog will be disabled.
-
-.. _modlog-command-modlogset-resetcases:
-
-""""""""""""""""""""
-modlogset resetcases
-""""""""""""""""""""
-
-**Syntax**
-
-.. code-block:: none
-
-    [p]modlogset resetcases 
-
-**Description**
-
-Reset all modlog cases in this server.
 
 .. _modlog-command-reason:
