+++ docs/index.rst
@@ -35,7 +35,6 @@ Welcome to Red - Discord Bot's documentation!
     intents
     cog_guides/admin
     cog_guides/alias
-    cog_guides/bank
     cog_guides/cleanup
     cog_guides/cog_manager_ui
     cog_guides/core
