--- redbot/cogs/bank/locales/de-DE.po
@@ -1,112 +1,0 @@
-msgid ""
-msgstr ""
-"Project-Id-Version: red-discordbot\n"
-"POT-Creation-Date: 2021-02-11 12:29+0000\n"
-"Last-Translator: \n"
-"Language-Team: German\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: redgettext 3.3\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-"X-Crowdin-Project: red-discordbot\n"
-"X-Crowdin-Project-ID: 289505\n"
-"X-Crowdin-Language: de\n"
-"X-Crowdin-File-ID: 14\n"
-"Language: de_DE\n"
-
-#: redbot/cogs/bank/bank.py:36
-#, docstring
-msgid "Bank"
-msgstr "Bank"
-
-#: redbot/cogs/bank/bank.py:48
-#, docstring
-msgid "Base command for bank settings."
-msgstr "Basis Befehl für Bank Einstellungen."
-
-#: redbot/cogs/bank/bank.py:52
-#, docstring
-msgid "Show the current bank settings."
-msgstr "Zeige die aktuellen Bank Einstellungen an."
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Global"
-msgstr "Überall"
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Server"
-msgstr "Server"
-
-#: redbot/cogs/bank/bank.py:67
-msgid "Bank settings:\n\n"
-"Bank name: {bank_name}\n"
-"Bank scope: {bank_scope}\n"
-"Currency: {currency_name}\n"
-"Default balance: {default_balance}\n"
-"Maximum allowed balance: {maximum_bal}\n"
-msgstr "Bankeinstellungen:\n\n"
-"Bankname: {bank_name}\n"
-"Bank bereich: {bank_scope}\n"
-"Währung: {currency_name}\n"
-"Standardbilanze: {default_balance}\n"
-"Maximal zulässiger Bilanzen: {maximum_bal}\n"
-
-#: redbot/cogs/bank/bank.py:83
-#, docstring
-msgid "Toggle whether the bank is global or not.\n\n"
-"        If the bank is global, it will become per-server.\n"
-"        If the bank is per-server, it will become global.\n"
-"        "
-msgstr "Schalte ein, ob die Bank global ist oder nicht.\n\n"
-"        Wenn die Bank global ist, wird sie pro Server.\n"
-"        Wenn die Bank pro Server ist, wird sie global.\n"
-"        "
-
-#: redbot/cogs/bank/bank.py:90
-msgid "per-server"
-msgstr "per-server"
-
-#: redbot/cogs/bank/bank.py:90
-msgid "global"
-msgstr "Global"
-
-#: redbot/cogs/bank/bank.py:93
-msgid "This will toggle the bank to be {banktype}, deleting all accounts in the process! If you're sure, type `{command}`"
-msgstr "Dies schaltet die Bank um zu {banktype}, löscht alle Konten im Prozess! Wenn du sicher bist, tippe `{command}`"
-
-#: redbot/cogs/bank/bank.py:100
-msgid "The bank is now {banktype}."
-msgstr "Die Bank ist nun {banktype}."
-
-#: redbot/cogs/bank/bank.py:106
-#, docstring
-msgid "Set the bank's name."
-msgstr "Legt den Namen der Bank fest."
-
-#: redbot/cogs/bank/bank.py:108
-msgid "Bank name has been set to: {name}"
-msgstr "Der Name der Bank wurde festgesetzt auf: {name}"
-
-#: redbot/cogs/bank/bank.py:114
-#, docstring
-msgid "Set the name for the bank's currency."
-msgstr "Lege den Namen für die Währung der Bank fest."
-
-#: redbot/cogs/bank/bank.py:116
-msgid "Currency name has been set to: {name}"
-msgstr "Währungs Name wurde gesetzt auf: {name}"
-
-#: redbot/cogs/bank/bank.py:122
-#, docstring
-msgid "Set the maximum balance a user can get."
-msgstr "Setze das maximale Guthaben fest, das ein Benutzer erhalten kann."
-
-#: redbot/cogs/bank/bank.py:128
-msgid "Amount must be greater than zero and less than {max}."
-msgstr "Der Betrag muss größer als null und kleiner als {max} sein."
-
-#: redbot/cogs/bank/bank.py:133
-msgid "Maximum balance has been set to: {amount}"
-msgstr "Das maximale Guthaben wurde festgelegt auf:{amount}"
-
