--- redbot/cogs/bank/locales/es-ES.po
@@ -1,112 +1,0 @@
-msgid ""
-msgstr ""
-"Project-Id-Version: red-discordbot\n"
-"POT-Creation-Date: 2021-02-11 12:29+0000\n"
-"Last-Translator: \n"
-"Language-Team: Spanish\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: redgettext 3.3\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-"X-Crowdin-Project: red-discordbot\n"
-"X-Crowdin-Project-ID: 289505\n"
-"X-Crowdin-Language: es-ES\n"
-"X-Crowdin-File-ID: 14\n"
-"Language: es_ES\n"
-
-#: redbot/cogs/bank/bank.py:36
-#, docstring
-msgid "Bank"
-msgstr "Banco"
-
-#: redbot/cogs/bank/bank.py:48
-#, docstring
-msgid "Base command for bank settings."
-msgstr "Comando básico para la configuración del banco."
-
-#: redbot/cogs/bank/bank.py:52
-#, docstring
-msgid "Show the current bank settings."
-msgstr "Muestra la configuración bancaria actual."
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Global"
-msgstr "Global"
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Server"
-msgstr "Servidor"
-
-#: redbot/cogs/bank/bank.py:67
-msgid "Bank settings:\n\n"
-"Bank name: {bank_name}\n"
-"Bank scope: {bank_scope}\n"
-"Currency: {currency_name}\n"
-"Default balance: {default_balance}\n"
-"Maximum allowed balance: {maximum_bal}\n"
-msgstr "Configuración de bancos:\n\n"
-"Nombre del banco: {bank_name}\n"
-"Alcance del banco: {bank_scope}\n"
-"Moneda: {currency_name}\n"
-"Saldo Predeterminado: {default_balance}\n"
-"Saldo Máximo Permitido: {maximum_bal}\n"
-
-#: redbot/cogs/bank/bank.py:83
-#, docstring
-msgid "Toggle whether the bank is global or not.\n\n"
-"        If the bank is global, it will become per-server.\n"
-"        If the bank is per-server, it will become global.\n"
-"        "
-msgstr "Alternar si el banco es global o no.\n\n"
-"        Si el banco es global, será por servidor.\n"
-"        Si el banco es por servidor, se volverá global.\n"
-"        "
-
-#: redbot/cogs/bank/bank.py:90
-msgid "per-server"
-msgstr "por server"
-
-#: redbot/cogs/bank/bank.py:90
-msgid "global"
-msgstr "general"
-
-#: redbot/cogs/bank/bank.py:93
-msgid "This will toggle the bank to be {banktype}, deleting all accounts in the process! If you're sure, type `{command}`"
-msgstr "Esto cambiará el banco a {banktype}, eliminando todas las cuentas en el proceso. Si está seguro, escriba `{command}`"
-
-#: redbot/cogs/bank/bank.py:100
-msgid "The bank is now {banktype}."
-msgstr "Este banco es ahora {banktype}."
-
-#: redbot/cogs/bank/bank.py:106
-#, docstring
-msgid "Set the bank's name."
-msgstr "Establece el nombre del banco."
-
-#: redbot/cogs/bank/bank.py:108
-msgid "Bank name has been set to: {name}"
-msgstr "Nombre del banco cambiado a: {name}"
-
-#: redbot/cogs/bank/bank.py:114
-#, docstring
-msgid "Set the name for the bank's currency."
-msgstr "Establezca el nombre de la moneda del banco."
-
-#: redbot/cogs/bank/bank.py:116
-msgid "Currency name has been set to: {name}"
-msgstr "La nombre de la moneda se ha establecido en: {name}"
-
-#: redbot/cogs/bank/bank.py:122
-#, docstring
-msgid "Set the maximum balance a user can get."
-msgstr "Establezca el equilibrio máximo que un usuario puede obtener."
-
-#: redbot/cogs/bank/bank.py:128
-msgid "Amount must be greater than zero and less than {max}."
-msgstr "La cantidad debe ser mayor que cero e menor que {max}."
-
-#: redbot/cogs/bank/bank.py:133
-msgid "Maximum balance has been set to: {amount}"
-msgstr "El saldo máximo se ha establecido en: {amount}"
-
