--- redbot/cogs/bank/locales/fr-FR.po
@@ -1,112 +1,0 @@
-msgid ""
-msgstr ""
-"Project-Id-Version: red-discordbot\n"
-"POT-Creation-Date: 2021-02-11 12:29+0000\n"
-"Last-Translator: \n"
-"Language-Team: French\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: redgettext 3.3\n"
-"Plural-Forms: nplurals=2; plural=(n > 1);\n"
-"X-Crowdin-Project: red-discordbot\n"
-"X-Crowdin-Project-ID: 289505\n"
-"X-Crowdin-Language: fr\n"
-"X-Crowdin-File-ID: 14\n"
-"Language: fr_FR\n"
-
-#: redbot/cogs/bank/bank.py:36
-#, docstring
-msgid "Bank"
-msgstr "Banque"
-
-#: redbot/cogs/bank/bank.py:48
-#, docstring
-msgid "Base command for bank settings."
-msgstr "Commande de base pour les paramètres de la banque."
-
-#: redbot/cogs/bank/bank.py:52
-#, docstring
-msgid "Show the current bank settings."
-msgstr "Affiche les paramètres actuels de la banque."
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Global"
-msgstr "Global"
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Server"
-msgstr "Serveur"
-
-#: redbot/cogs/bank/bank.py:67
-msgid "Bank settings:\n\n"
-"Bank name: {bank_name}\n"
-"Bank scope: {bank_scope}\n"
-"Currency: {currency_name}\n"
-"Default balance: {default_balance}\n"
-"Maximum allowed balance: {maximum_bal}\n"
-msgstr "Paramètres de la banque :\n\n"
-"Nom de la banque : {bank_name}\n"
-"Contexte de la banque : {bank_scope}\n"
-"Devise : {currency_name}\n"
-"Solde de départ : {default_balance}\n"
-"Solde maximum autorisé : {maximum_bal}\n"
-
-#: redbot/cogs/bank/bank.py:83
-#, docstring
-msgid "Toggle whether the bank is global or not.\n\n"
-"        If the bank is global, it will become per-server.\n"
-"        If the bank is per-server, it will become global.\n"
-"        "
-msgstr "Permet de choisir si la banque est globale ou non.\n\n"
-"        Si la banque est globale, elle deviendra par serveur.\n"
-"        Si la banque est par serveurs, elle deviendra globale.\n"
-"        "
-
-#: redbot/cogs/bank/bank.py:90
-msgid "per-server"
-msgstr "par serveur"
-
-#: redbot/cogs/bank/bank.py:90
-msgid "global"
-msgstr "global"
-
-#: redbot/cogs/bank/bank.py:93
-msgid "This will toggle the bank to be {banktype}, deleting all accounts in the process! If you're sure, type `{command}`"
-msgstr "Cela va basculer la banque en : {banktype}, en supprimant tout les comptes durant le processus ! Si vous êtes sûr, tapez `{command}`"
-
-#: redbot/cogs/bank/bank.py:100
-msgid "The bank is now {banktype}."
-msgstr "La banque est désormais : {banktype}."
-
-#: redbot/cogs/bank/bank.py:106
-#, docstring
-msgid "Set the bank's name."
-msgstr "Défini le nom de la banque."
-
-#: redbot/cogs/bank/bank.py:108
-msgid "Bank name has been set to: {name}"
-msgstr "Le nom de la banque est maintenant : {name}"
-
-#: redbot/cogs/bank/bank.py:114
-#, docstring
-msgid "Set the name for the bank's currency."
-msgstr "Défini le nom de la devise de la banque."
-
-#: redbot/cogs/bank/bank.py:116
-msgid "Currency name has been set to: {name}"
-msgstr "Le nom de la devise est maintenant : {name}"
-
-#: redbot/cogs/bank/bank.py:122
-#, docstring
-msgid "Set the maximum balance a user can get."
-msgstr "Définit le solde maximum qu'un utilisateur peut obtenir."
-
-#: redbot/cogs/bank/bank.py:128
-msgid "Amount must be greater than zero and less than {max}."
-msgstr "Le montant doit être supérieur à zéro et inférieur à {max}."
-
-#: redbot/cogs/bank/bank.py:133
-msgid "Maximum balance has been set to: {amount}"
-msgstr "Le solde maximal a été réglé sur : {amount}"
-
