--- redbot/cogs/bank/locales/lol-US.po
@@ -1,71 +1,0 @@
-msgid ""
-msgstr ""
-"Project-Id-Version: red-discordbot\n"
-"POT-Creation-Date: 2019-07-05 22:33+0200\n"
-"PO-Revision-Date: 2019-07-06 16:14\n"
-"Last-Translator: Robert Jansen (Kowlin)\n"
-"Language-Team: LOLCAT\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: redgettext 3.0\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-"X-Generator: crowdin.com\n"
-"X-Crowdin-Project: red-discordbot\n"
-"X-Crowdin-Language: lol\n"
-"X-Crowdin-File: /cogs/bank/locales/messages.pot\n"
-"Language: lol_US\n"
-
-#: redbot/cogs/bank/bank.py:62
-#, docstring
-msgid "Bank"
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:74
-#, docstring
-msgid "Base command for bank settings."
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:87
-msgid "Bank settings:\\n\\nBank name: {bank_name}\\nCurrency: {currency_name}\\nDefault balance: {default_balance}"
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:98
-#, docstring
-msgid "Toggle whether the bank is global or not.\\n\\n        If the bank is global, it will become per-server.\\n        If the bank is per-server, it will become global.\\n        "
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:105
-msgid "per-server"
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:105
-msgid "global"
-msgstr "global"
-
-#: redbot/cogs/bank/bank.py:108
-msgid "This will toggle the bank to be {banktype}, deleting all accounts in the process! If you're sure, type `{command}`"
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:115
-msgid "The bank is now {banktype}."
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:120
-#, docstring
-msgid "Set the bank's name."
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:122
-msgid "Bank name has been set to: {name}"
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:127
-#, docstring
-msgid "Set the name for the bank's currency."
-msgstr ""
-
-#: redbot/cogs/bank/bank.py:129
-msgid "Currency name has been set to: {name}"
-msgstr ""
-
