--- redbot/cogs/bank/locales/pl-PL.po
@@ -1,112 +1,0 @@
-msgid ""
-msgstr ""
-"Project-Id-Version: red-discordbot\n"
-"POT-Creation-Date: 2021-02-11 12:29+0000\n"
-"Last-Translator: \n"
-"Language-Team: Polish\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: redgettext 3.3\n"
-"Plural-Forms: nplurals=4; plural=(n==1 ? 0 : (n%10>=2 && n%10<=4) && (n%100<12 || n%100>14) ? 1 : n!=1 && (n%10>=0 && n%10<=1) || (n%10>=5 && n%10<=9) || (n%100>=12 && n%100<=14) ? 2 : 3);\n"
-"X-Crowdin-Project: red-discordbot\n"
-"X-Crowdin-Project-ID: 289505\n"
-"X-Crowdin-Language: pl\n"
-"X-Crowdin-File-ID: 14\n"
-"Language: pl_PL\n"
-
-#: redbot/cogs/bank/bank.py:36
-#, docstring
-msgid "Bank"
-msgstr "Bank"
-
-#: redbot/cogs/bank/bank.py:48
-#, docstring
-msgid "Base command for bank settings."
-msgstr "Podstawowa komenda do ustawień banku."
-
-#: redbot/cogs/bank/bank.py:52
-#, docstring
-msgid "Show the current bank settings."
-msgstr "Pokazuje bieżące ustawienia banku."
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Global"
-msgstr "Globalny"
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Server"
-msgstr "Serwer"
-
-#: redbot/cogs/bank/bank.py:67
-msgid "Bank settings:\n\n"
-"Bank name: {bank_name}\n"
-"Bank scope: {bank_scope}\n"
-"Currency: {currency_name}\n"
-"Default balance: {default_balance}\n"
-"Maximum allowed balance: {maximum_bal}\n"
-msgstr "Ustawienia banku:\n\n"
-"Nazwa banku: {bank_name}\n"
-"Zakres banku: {bank_scope}\n"
-"Waluta: {currency_name}\n"
-"Domyślne saldo: {default_balance}\n"
-"Maksymalne dozwolone saldo: {maximum_bal}\n"
-
-#: redbot/cogs/bank/bank.py:83
-#, docstring
-msgid "Toggle whether the bank is global or not.\n\n"
-"        If the bank is global, it will become per-server.\n"
-"        If the bank is per-server, it will become global.\n"
-"        "
-msgstr "Przełącza ustawienie, czy bank jest globalny, czy nie.\n\n"
-"        Jeśli bank jest globalny, będzie przypisany do poszczególnych serwerów.\n"
-"        Jeśli bank jest przeznaczony na serwer, stanie się globalny.\n"
-"        "
-
-#: redbot/cogs/bank/bank.py:90
-msgid "per-server"
-msgstr "per-server"
-
-#: redbot/cogs/bank/bank.py:90
-msgid "global"
-msgstr "globalny"
-
-#: redbot/cogs/bank/bank.py:93
-msgid "This will toggle the bank to be {banktype}, deleting all accounts in the process! If you're sure, type `{command}`"
-msgstr "Spowoduje to przełączenie banku na {banktype} i usunięcie wszystkich kont w wyniku tego procesu! Jeśli jesteś pewien/pewna, wpisz `{command}`"
-
-#: redbot/cogs/bank/bank.py:100
-msgid "The bank is now {banktype}."
-msgstr "Bank jest teraz {banktype}."
-
-#: redbot/cogs/bank/bank.py:106
-#, docstring
-msgid "Set the bank's name."
-msgstr "Ustaw nazwę banku."
-
-#: redbot/cogs/bank/bank.py:108
-msgid "Bank name has been set to: {name}"
-msgstr "Nazwa banku została ustawiona na: {name}"
-
-#: redbot/cogs/bank/bank.py:114
-#, docstring
-msgid "Set the name for the bank's currency."
-msgstr "Ustawia nazwę waluty banku."
-
-#: redbot/cogs/bank/bank.py:116
-msgid "Currency name has been set to: {name}"
-msgstr "Nazwa waluty została ustawiona na: {name}"
-
-#: redbot/cogs/bank/bank.py:122
-#, docstring
-msgid "Set the maximum balance a user can get."
-msgstr "Ustawia maksymalne saldo, jakie może uzyskać użytkownik."
-
-#: redbot/cogs/bank/bank.py:128
-msgid "Amount must be greater than zero and less than {max}."
-msgstr "Kwota musi być większa od zera i mniejsza niż {max}."
-
-#: redbot/cogs/bank/bank.py:133
-msgid "Maximum balance has been set to: {amount}"
-msgstr "Maksymalne saldo zostało ustawione na: {amount}"
-
