--- redbot/cogs/bank/locales/sl-SI.po
@@ -1,112 +1,0 @@
-msgid ""
-msgstr ""
-"Project-Id-Version: red-discordbot\n"
-"POT-Creation-Date: 2021-02-11 12:29+0000\n"
-"Last-Translator: \n"
-"Language-Team: Slovenian\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: redgettext 3.3\n"
-"Plural-Forms: nplurals=4; plural=(n%100==1 ? 1 : n%100==2 ? 2 : n%100==3 || n%100==4 ? 3 : 0);\n"
-"X-Crowdin-Project: red-discordbot\n"
-"X-Crowdin-Project-ID: 289505\n"
-"X-Crowdin-Language: sl\n"
-"X-Crowdin-File-ID: 14\n"
-"Language: sl_SI\n"
-
-#: redbot/cogs/bank/bank.py:36
-#, docstring
-msgid "Bank"
-msgstr "Banka"
-
-#: redbot/cogs/bank/bank.py:48
-#, docstring
-msgid "Base command for bank settings."
-msgstr "Osnovni ukazi za upravljanje nastavitev banke."
-
-#: redbot/cogs/bank/bank.py:52
-#, docstring
-msgid "Show the current bank settings."
-msgstr "Prikaže trenutne nastavitve banke."
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Global"
-msgstr "Globalno"
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Server"
-msgstr "Strežnik"
-
-#: redbot/cogs/bank/bank.py:67
-msgid "Bank settings:\n\n"
-"Bank name: {bank_name}\n"
-"Bank scope: {bank_scope}\n"
-"Currency: {currency_name}\n"
-"Default balance: {default_balance}\n"
-"Maximum allowed balance: {maximum_bal}\n"
-msgstr "Nastavitve banke:\n\n"
-"Ime banke: {bank_name}\n"
-"Obseg banke: {bank_scope}\n"
-"Valuta: {currency_name}\n"
-"Privzeto stanje: {default_balance}\n"
-"Maksimalno dovoljeno stanje: {maximum_bal}\n"
-
-#: redbot/cogs/bank/bank.py:83
-#, docstring
-msgid "Toggle whether the bank is global or not.\n\n"
-"        If the bank is global, it will become per-server.\n"
-"        If the bank is per-server, it will become global.\n"
-"        "
-msgstr "Preklopite, ali je banka globalna ali ne.\n\n"
-"        Če je banka globalna, bo zdaj omejena na vsak strežnik posebej.\n"
-"        Če je banka na vsak strežnik posebej, bo zdaj globalna.\n"
-"        "
-
-#: redbot/cogs/bank/bank.py:90
-msgid "per-server"
-msgstr "na strežnik"
-
-#: redbot/cogs/bank/bank.py:90
-msgid "global"
-msgstr "globalno"
-
-#: redbot/cogs/bank/bank.py:93
-msgid "This will toggle the bank to be {banktype}, deleting all accounts in the process! If you're sure, type `{command}`"
-msgstr "To bo preklopilo banko na vrsto {banktype}, kar bo izbrisalo vse račune med procesom! Če si prepričan uporabi ukaz `{command}`"
-
-#: redbot/cogs/bank/bank.py:100
-msgid "The bank is now {banktype}."
-msgstr "Banka je zdaj {banktype}."
-
-#: redbot/cogs/bank/bank.py:106
-#, docstring
-msgid "Set the bank's name."
-msgstr "Nastavi ime banke."
-
-#: redbot/cogs/bank/bank.py:108
-msgid "Bank name has been set to: {name}"
-msgstr "Ime banke je zdaj {name}"
-
-#: redbot/cogs/bank/bank.py:114
-#, docstring
-msgid "Set the name for the bank's currency."
-msgstr "Nastavi ime valute za banko."
-
-#: redbot/cogs/bank/bank.py:116
-msgid "Currency name has been set to: {name}"
-msgstr "Ime valute je zdaj {name}"
-
-#: redbot/cogs/bank/bank.py:122
-#, docstring
-msgid "Set the maximum balance a user can get."
-msgstr "Nastavi maksimalno stanje katerega lahko uporabnik doseže."
-
-#: redbot/cogs/bank/bank.py:128
-msgid "Amount must be greater than zero and less than {max}."
-msgstr "Znesek mora biti večji od nič in manjši od {max}."
-
-#: redbot/cogs/bank/bank.py:133
-msgid "Maximum balance has been set to: {amount}"
-msgstr "Maksimalno stanje je nastavljeno na {amount}"
-
