--- redbot/cogs/bank/locales/tr-TR.po
@@ -1,112 +1,0 @@
-msgid ""
-msgstr ""
-"Project-Id-Version: red-discordbot\n"
-"POT-Creation-Date: 2021-02-11 12:29+0000\n"
-"Last-Translator: \n"
-"Language-Team: Turkish\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: redgettext 3.3\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-"X-Crowdin-Project: red-discordbot\n"
-"X-Crowdin-Project-ID: 289505\n"
-"X-Crowdin-Language: tr\n"
-"X-Crowdin-File-ID: 14\n"
-"Language: tr_TR\n"
-
-#: redbot/cogs/bank/bank.py:36
-#, docstring
-msgid "Bank"
-msgstr "Banka"
-
-#: redbot/cogs/bank/bank.py:48
-#, docstring
-msgid "Base command for bank settings."
-msgstr "Banka ayarları için temel komut."
-
-#: redbot/cogs/bank/bank.py:52
-#, docstring
-msgid "Show the current bank settings."
-msgstr "Bankanın şu anki ayarlarını göster."
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Global"
-msgstr "Küresel"
-
-#: redbot/cogs/bank/bank.py:62
-msgid "Server"
-msgstr "Sunucu"
-
-#: redbot/cogs/bank/bank.py:67
-msgid "Bank settings:\n\n"
-"Bank name: {bank_name}\n"
-"Bank scope: {bank_scope}\n"
-"Currency: {currency_name}\n"
-"Default balance: {default_balance}\n"
-"Maximum allowed balance: {maximum_bal}\n"
-msgstr "Banka ayarları:\n\n"
-"Banka ismi: {bank_name}\n"
-"Banka kapsamı: {bank_scope}\n"
-"Para birimi: {currency_name}\n"
-"Varsayılan bakiye: {default_balance}\n"
-"Maksimum izin verilen bakiye: {maximum_bal}\n"
-
-#: redbot/cogs/bank/bank.py:83
-#, docstring
-msgid "Toggle whether the bank is global or not.\n\n"
-"        If the bank is global, it will become per-server.\n"
-"        If the bank is per-server, it will become global.\n"
-"        "
-msgstr "Banka küresel olsun veya olmasın aç/kapa.\n\n"
-"        Eğer banka küresel ise, sunuculara özele çevrilir.\n"
-"        Eğer banka sunuculara özelse, küresele çevrilir.\n"
-"        "
-
-#: redbot/cogs/bank/bank.py:90
-msgid "per-server"
-msgstr "sunucu başına"
-
-#: redbot/cogs/bank/bank.py:90
-msgid "global"
-msgstr "genel"
-
-#: redbot/cogs/bank/bank.py:93
-msgid "This will toggle the bank to be {banktype}, deleting all accounts in the process! If you're sure, type `{command}`"
-msgstr "Bu bankanın işleminin tüm hesapları silme {banktype}, geçiş yapılır! Emin misin, '{command} ' yazın"
-
-#: redbot/cogs/bank/bank.py:100
-msgid "The bank is now {banktype}."
-msgstr "Banka şimdi {banktype}."
-
-#: redbot/cogs/bank/bank.py:106
-#, docstring
-msgid "Set the bank's name."
-msgstr "Bankanın adını ayarlayın."
-
-#: redbot/cogs/bank/bank.py:108
-msgid "Bank name has been set to: {name}"
-msgstr "Banka adı ayarlandı {name}"
-
-#: redbot/cogs/bank/bank.py:114
-#, docstring
-msgid "Set the name for the bank's currency."
-msgstr "Bankası'nın para birimi için ayarlayın."
-
-#: redbot/cogs/bank/bank.py:116
-msgid "Currency name has been set to: {name}"
-msgstr "Para birimi adı ayarla: {name}"
-
-#: redbot/cogs/bank/bank.py:122
-#, docstring
-msgid "Set the maximum balance a user can get."
-msgstr "Bir kullanıcının alabileceği maksimum bakiyeyi ayarlayın."
-
-#: redbot/cogs/bank/bank.py:128
-msgid "Amount must be greater than zero and less than {max}."
-msgstr "Bu sayı 0 dan yüksek ve {max} bu sayıdan az olmalı."
-
-#: redbot/cogs/bank/bank.py:133
-msgid "Maximum balance has been set to: {amount}"
-msgstr "Maksimum bakiye şu şekilde ayarlandı: {amount}"
-
