+++ redbot/cogs/cleanup/cleanup.py
@@ -5,9 +5,9 @@ from typing import Callable, List, Optional, Set, Union
 
 import discord
 
-from redbot.cogs.mod.converters import RawUserIds
 from redbot.core import checks, commands, Config
 from redbot.core.bot import Red
+from redbot.core.commands import RawUserIdConverter
 from redbot.core.i18n import Translator, cog_i18n
 from redbot.core.utils.chat_formatting import humanize_number
 from redbot.core.utils.mod import slow_deletion, mass_purge
@@ -232,7 +232,7 @@ class Cleanup(commands.Cog):
     async def user(
         self,
         ctx: commands.Context,
-        user: Union[discord.Member, RawUserIds],
+        user: Union[discord.Member, RawUserIdConverter],
         number: positive_int,
         delete_pinned: bool = False,
     ):
