+++ redbot/core/bank.py
@@ -43,6 +43,7 @@ __all__ = [
     "cost",
     "AbortPurchase",
     "bank_prune",
+    "is_owner_if_bank_global",
 ]
 
 _MAX_BALANCE = 2 ** 63 - 1
@@ -145,6 +146,44 @@ async def _process_data_deletion(
         async for guild_id, member_dict in AsyncIter(all_members.items(), steps=100):
             if user_id in member_dict:
                 await _config.member_from_ids(guild_id, user_id).clear()
+
+
+def is_owner_if_bank_global():
+    """
+    Restrict the command to the bot owner if the bank is global,
+    otherwise ensure it's used in guild (WITHOUT checking any user permissions).
+
+    When used on the command, this should be combined
+    with permissions check like `guildowner_or_permissions()`.
+
+    This is a `command check <discord.ext.commands.check>`.
+
+    Example
+    -------
+
+    .. code-block:: python
+
+        @bank.is_owner_if_bank_global()
+        @checks.guildowner()
+        @commands.group()
+        async def bankset(self, ctx: commands.Context):
+            \"""Base command for bank settings.\"""
+
+    If the bank is global, the ``[p]bankset`` command can only be used by
+    the bot owners in both guilds and DMs.
+    If the bank is local, the command can only be used in guilds by guild and bot owners.
+    """
+
+    async def pred(ctx: commands.Context):
+        author = ctx.author
+        if not await is_global():
+            if not ctx.guild:
+                return False
+            return True
+        else:
+            return await ctx.bot.is_owner(author)
+
+    return commands.check(pred)
 
 
 class Account:
