+++ redbot/core/commands/__init__.py
@@ -30,6 +30,7 @@ from .converter import (
     NoParseOptional as NoParseOptional,
     UserInputOptional as UserInputOptional,
     Literal as Literal,
+    RawUserIdConverter as RawUserIdConverter,
     CogConverter as CogConverter,
     CommandConverter as CommandConverter,
 )
