+++ redbot/core/commands/converter.py
@@ -34,6 +34,7 @@ if TYPE_CHECKING:
     from .context import Context
 
 __all__ = [
+    "RawUserIdConverter",
     "DictConverter",
     "UserInputOptional",
     "NoParseOptional",
@@ -51,6 +52,7 @@ __all__ = [
 _ = Translator("commands.converter", __file__)
 
 ID_REGEX = re.compile(r"([0-9]{15,20})")
+USER_MENTION_REGEX = re.compile(r"<@!?([0-9]{15,21})>$")
 
 
 # Taken with permission from
@@ -202,6 +204,32 @@ def parse_relativedelta(
             )
         return delta
     return None
+
+
+class RawUserIdConverter(dpy_commands.Converter):
+    """
+    Converts ID or user mention to an `int`.
+
+    Useful for commands like ``[p]ban`` or ``[p]unban`` where the bot is not necessarily
+    going to share any servers with the user that a moderator wants to ban/unban.
+
+    This converter doesn't check if the ID/mention points to an actual user
+    but it won't match IDs and mentions that couldn't possibly be valid.
+
+    For example, the converter will not match on "123" because the number doesn't have
+    enough digits to be valid ID but, it will match on "12345678901234567" even though
+    there is no user with such ID.
+    """
+
+    async def convert(self, ctx: "Context", argument: str) -> int:
+        # This is for the hackban and unban commands, where we receive IDs that
+        # are most likely not in the guild.
+        # Mentions are supported, but most likely won't ever be in cache.
+
+        if match := ID_REGEX.match(argument) or USER_MENTION_REGEX.match(argument):
+            return int(match.group(1))
+
+        raise BadArgument(_("'{input}' doesn't look like a valid user ID.").format(input=argument))
 
 
 # Below this line are a lot of lies for mypy about things that *end up* correct when
